chapter 8 stuff

+++ average4.py
@@ -4,7 +4,7 @@ def main():
     arr = []
     while True:
         try: 
-            arr.append(int(input("Enter a number :")))
+            arr.append(float(input("Enter a number :")))
         except ValueError:
             break
     print("Average:",sum(arr)/len(arr))
